fix: Renamed Services directory to services in shared-lib and teachers app to resolve case-sensitive build failures on Linux/Vercel
