+++ libs/shared-lib/src/index.ts
@@ -12,10 +12,10 @@ export * from './lib/Tabs/CommonTabs';
 export * from './lib/Card/CommonCard';
 export * from './lib/Collapse/CommonAccordian';
 export * from './lib/Drawer/CommonDrawer';
-export * from './lib/Services/LoginService';
-export * from './lib/Services/RestClient';
-export * from './lib/Services/Interceptor';
-export * from './lib/Services/Content/Search';
+export * from './lib/services/LoginService';
+export * from './lib/services/RestClient';
+export * from './lib/services/Interceptor';
+export * from './lib/services/Content/Search';
 export * from './lib/SunbirdPlayer/SunbirdPlayer';
 export * from './lib/QuestionSetEditor/QuestionSet';
 export * from './lib/CollectionEditor/Collection';
