+++ libs/shared-lib/src/lib/services/app.constant.ts
